change in controllers development of dto and mapper

+++ src/main/java/project/financement/service/ExpenseCategoryService.java
@@ -3,9 +3,13 @@ package project.financement.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import project.financement.dto.ExpenseCategoryDto;
 import project.financement.entity.ExpenseCategory;
+import project.financement.exception.ExpenseCategoryDeletionException;
 import project.financement.exception.ExpenseCategoryNotFoundException;
+import project.financement.mapper.ExpenseCategoryMapper;
 import project.financement.repository.ExpenseCategoryRepository;
+import project.financement.repository.ExpenseRepository;
 
 import java.util.List;
 import java.util.UUID;
@@ -14,30 +18,50 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class ExpenseCategoryService {
     private final ExpenseCategoryRepository expenseCategoryRepository;
+    private final ExpenseRepository expenseRepository;
+    private final ExpenseCategoryMapper expenseCategoryMapper;
 
-    public List<ExpenseCategory> findAll() {
-        return expenseCategoryRepository.findAll();
-    }
-
-    public ExpenseCategory findExpenseCategoryById(UUID id) {
-        return expenseCategoryRepository.findById(id).orElseThrow(() ->
-                new ExpenseCategoryNotFoundException(id));
+    @Transactional
+    public List<ExpenseCategoryDto> findAll() {
+        List<ExpenseCategory> expenseCategories = expenseCategoryRepository.findAll();
+        return expenseCategoryMapper.toDto(expenseCategories);
     }
 
     @Transactional
-    public ExpenseCategory saveExpenseCategory(ExpenseCategory expenseCategory) {
-        return expenseCategoryRepository.save(expenseCategory);
+    public ExpenseCategoryDto findExpenseCategoryById(UUID id) {
+        ExpenseCategory expenseCategory = expenseCategoryRepository.findById(id).orElseThrow(() ->
+                new ExpenseCategoryNotFoundException(id));
+        return expenseCategoryMapper.toDto(expenseCategory);
+    }
+
+    @Transactional
+    public ExpenseCategoryDto findByExpenseCategoryName(String expenseCategoryName) {
+        ExpenseCategory expenseCategory = expenseCategoryRepository.findByExpenseCategoryName(expenseCategoryName)
+                .orElseThrow(() -> new ExpenseCategoryNotFoundException("Expense category not found"));
+        return expenseCategoryMapper.toDto(expenseCategory);
+    }
+
+    @Transactional
+    public ExpenseCategoryDto saveExpenseCategory(ExpenseCategoryDto expenseCategoryDto) {
+        ExpenseCategory expenseCategory = expenseCategoryMapper.toEntity(expenseCategoryDto);
+        ExpenseCategory savedExpenseCategory = expenseCategoryRepository.save(expenseCategory);
+        return expenseCategoryMapper.toDto(savedExpenseCategory);
+    }
+
+    @Transactional
+    public ExpenseCategoryDto updateExpenseCategory(UUID id, ExpenseCategoryDto expenseCategoryDto) {
+        ExpenseCategory expenseCategoryToUpdate = expenseCategoryRepository.findById(id)
+                .orElseThrow(() -> new ExpenseCategoryNotFoundException(id));
+        expenseCategoryToUpdate.setExpenseCategoryName(expenseCategoryDto.getExpenseCategoryName());
+        ExpenseCategory updatedExpenseCategory = expenseCategoryRepository.save(expenseCategoryToUpdate);
+        return expenseCategoryMapper.toDto(updatedExpenseCategory);
     }
 
     @Transactional
     public void deleteExpenseCategoryById(UUID id) {
+        if (expenseRepository.existsByExpenseCategoryName_ExpenseCategoryId(id)) {
+            throw new ExpenseCategoryDeletionException("Cannot delete expense category because it is referenced by one or more expenses.");
+        }
         expenseCategoryRepository.deleteById(id);
-    }
-
-    @Transactional
-    public ExpenseCategory updateExpenseCategory(UUID id, ExpenseCategory expenseCategory) {
-        ExpenseCategory expenseCategoryToUpdate = findExpenseCategoryById(id);
-        expenseCategoryToUpdate.setExpenseCategoryName(expenseCategory.getExpenseCategoryName());
-        return expenseCategoryRepository.save(expenseCategoryToUpdate);
     }
 }
